historical-load/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-functions
pytest==5.4.3
pyarrow
